divide o arquivo de helperFunctions em mais arquivos na pasta util
